+++ src/de/omegazirkel/risingworld/template/ui/TemplatePlayerPluginData.java
@@ -1,0 +1,27 @@
+package de.omegazirkel.risingworld.template.ui;
+
+import de.omegazirkel.risingworld.MavenTemplate;
+import de.omegazirkel.risingworld.tools.ui.BasePlayerPluginDataPanel;
+import de.omegazirkel.risingworld.tools.ui.PlayerPluginData;
+import net.risingworld.api.objects.Player;
+
+/* This is an example, rename <Template> if you use this feature */
+
+public class TemplatePlayerPluginData extends PlayerPluginData {
+
+    public TemplatePlayerPluginData(String pluginVersion) {
+        this.pluginLabel = MavenTemplate.name;
+        this.pluginVersion = pluginVersion;
+    }
+
+    @Override
+    public BasePlayerPluginDataPanel createPlayerPluginDataUIElement(Player uiPlayer) {
+        return new BasePlayerPluginDataPanel(uiPlayer, pluginLabel) {
+            @Override
+            protected void redrawContent() {
+                flexWrapper.removeAllChilds();
+                flexWrapper.addChild(defaultEmptyStateLabel());
+            }
+        };
+    }
+}